Move get_detect_and_deblend_task to photometry.py
