+++ metadetect/lsst/measure.py
@@ -263,62 +263,6 @@ class DetectAndDeblendTask(Task):
         return sources, model_data
 
 
-def get_detect_and_deblend_task(
-    rng=None,
-    thresh=DEFAULT_THRESH,
-    deblender=None,
-    config=None,
-):
-    """
-    run detection and deblending of peaks, as well as basic measurments such as
-    centroid.  The SDSS deblender is run in order to split footprints.
-
-    We must combine detection and deblending in the same function because the
-    schema gets modified in place, which means we must construct the deblend
-    task at the same time as the detect task
-
-    Parameters
-    ----------
-    rng: np.random.RandomState
-        Random number generator for noise replacer
-    thresh: float, optional
-        The detection threshold in units of the sky noise
-    config: dict, optional
-        The configuration dictionary to override the defaults with.
-
-    Returns
-    -------
-    sources, detexp
-        The sources and the detection exposure
-    """
-    if deblender is not None:
-        LOG.warning(
-            "'deblender' kwargs is not used and will be removed soon. "
-            "Specify the deblender via the config kwarg instead."
-        )
-
-    config_override = config if config is not None else {}
-    if thresh:
-        if 'detect' not in config_override:
-            config_override['detect'] = {}
-        config_override['detect']['thresholdValue'] = thresh
-
-    config = DetectAndDeblendConfig()
-    config.setDefaults()
-
-    if config_override.get('deblend', {}).pop('name', '') == "scarlet":
-        config.deblend.retarget(ScarletDeblendTask)
-
-    util.override_config(config, config_override)
-
-    config.freeze()
-    config.validate()
-    task = DetectAndDeblendTask(config=config)
-    if rng is not None:
-        task.rng = rng
-    return task
-
-
 def measure(
     mbexp,
     detexp,
