+++ metadetect/lsst/tests/test_lsst_skysub.py
@@ -5,7 +5,7 @@ import pytest
 import tqdm
 import logging
 import metadetect.lsst.skysub as lsst_skysub
-import metadetect.lsst.measure as lsst_measure
+from metadetect.lsst.photometry import get_detect_and_deblend_task
 
 logging.basicConfig(
     stream=sys.stdout,
@@ -266,7 +266,7 @@ def test_skysub_sim_wldeblend_gal(star_density, sky_n_sigma):
         else:
             # this one is for debugging; we do the iterations ourselves so we
             # can display the result
-            dbtask = lsst_measure.get_detect_and_deblend_task(thresh=5)
+            dbtask = get_detect_and_deblend_task(thresh=5)
             _ = dbtask.run(exp)
             if False:
                 show_mask(exp)
